#120: fix edit page padding and add debounce to note add
